+++ runtime/msgpack/adaptor/class_instance.cpp
@@ -2,12 +2,8 @@
 // Copyright (c) 2020 LLC «V Kontakte»
 // Distributed under the GPL v3 License, see LICENSE.notice.txt
 
-#include "runtime/msgpack-serialization.h"
-
-#include "msgpack/adaptors.h"
+#include "runtime/msgpack/adaptor/class_instance.h"
 
 namespace msgpack::adaptor {
-
 size_t CheckInstanceDepth::depth = 0;
-
 } // namespace msgpack::adaptor
