+++ runtime/msgpack/adaptors.h
@@ -1,4 +1,9 @@
+#include "runtime/msgpack/adaptor/array.h"
 #include "runtime/msgpack/adaptor/bool.h"
+#include "runtime/msgpack/adaptor/class_instance.h"
 #include "runtime/msgpack/adaptor/float.h"
 #include "runtime/msgpack/adaptor/int.h"
+#include "runtime/msgpack/adaptor/mixed.h"
+#include "runtime/msgpack/adaptor/optional.h"
+#include "runtime/msgpack/adaptor/string.h"
 #include "runtime/msgpack/adaptor/tuple.h"
